extracted displaySelectedImageHandler helper
#20
